Add SymmetricOrPalindromo class
A new file, SymmetricOrPalindromo.java, has been added to the org.example.treiniprojectjava package. This class contains a method to check whether a given string is a palindrome, and a main method that takes user input and prints either "1" if the input is a palindrome, or a random number if it's not.